+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## Fixed
+
+- When creating new git remote name, check it does not already exist. If it does, add random numbers to name.
+
 ## [0.4.0] - 2022-11-18
 
 ## Added